requirements.txt
autopep8==1.5.6
numpy==1.20.1
Pillow==8.1.2
pycodestyle==2.7.0
toml==0.10.2
